Data-Engineering-7/api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
cassandra-driver==3.29.2
redis==5.0.7
python-dotenv==1.0.1
